requirements.txt
altgraph==0.17
EasyProcess==0.3
entrypoint2==0.2.3
future==0.18.2
MouseInfo==0.1.3
mss==6.1.0
pefile==2019.4.18
Pillow==8.1.0
PyAutoGUI==0.9.52
PyGetWindow==0.0.9
pyinstaller==4.2
pyinstaller-hooks-contrib==2020.11
PyMsgBox==1.0.9
pynput==1.6.8
pyperclip==1.8.1
pypiwin32==223
PyRect==0.1.4
pyscreenshot==2.3
PyScreeze==0.1.26
PyTweening==1.0.3
pywin32==300
pywin32-ctypes==0.2.0
six==1.15.0
